+++ src/sys/boot/multiboot.rs
@@ -16,10 +16,8 @@ static MULTIBOOT_HEADER: [u32; 6] = [
 // TODO: Improve protocol support
 pub extern "C" fn start(info: u32, magic: u32) -> ! {
     crate::sys::vga::init();
-    crate::sys::vga::print_fmt(format_args!("MOROS loading...\n"));
 
-    crate::sys::serial::init();
-    crate::sys::serial::print_fmt(format_args!("MOROS loading...\n"));
+    printk!("MOROS loading...\n");
 
     if magic == multiboot2::MAGIC {
         let boot_info = unsafe {
@@ -40,11 +38,14 @@ pub extern "C" fn start(info: u32, magic: u32) -> ! {
             //let offset = 0;
             //crate::init(&memory_map, offset);
             crate::sys::gdt::init();
+            crate::sys::idt::init();
+            crate::sys::pic::init();
+            crate::sys::x86::int::enable_interrupts();
+            crate::sys::serial::init();
         }
     }
 
-    crate::sys::serial::print_fmt(format_args!("MOROS loaded successfully!\n"));
-    crate::sys::vga::print_fmt(format_args!("MOROS loaded successfully!\n"));
+    printk!("MOROS loaded successfully!\n");
 
     //crate::exec();
     crate::hang();